+++ src/Proxy/ProxyAutoloaderInterface.php
@@ -11,11 +11,10 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-namespace Rekalogika\Mapper\Transformer\Proxy;
+namespace Rekalogika\Mapper\Proxy;
 
-interface ProxyRegistryInterface
+interface ProxyAutoloaderInterface
 {
-    public function registerProxy(
-        ProxySpecification $proxySpecification,
-    ): void;
+    public function registerAutoloader(): void;
+    public function unregisterAutoloader(): void;
 }