+++ src/Proxy/ProxyRegistryInterface.php
@@ -11,10 +11,11 @@ declare(strict_types=1);
  * that was distributed with this source code.
  */
 
-namespace Rekalogika\Mapper\Transformer\Proxy;
+namespace Rekalogika\Mapper\Proxy;
 
-interface ProxyAutoloaderInterface
+interface ProxyRegistryInterface
 {
-    public function registerAutoloader(): void;
-    public function unregisterAutoloader(): void;
+    public function registerProxy(
+        ProxySpecification $proxySpecification,
+    ): void;
 }